Refactoring to move interfaces and SubstringDescriptor into separate files
