+++ SilUtils/SilUtils.csproj
@@ -135,6 +135,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ListUtils.cs" />
     <Compile Include="StringUtils.cs" />
+    <Compile Include="SubstringDescriptor.cs" />
     <Compile Include="TextBoxExtensions.cs" />
     <Compile Include="WaitCursor.cs" />
     <Compile Include="SerializationHelper.cs" />
