+++ Transcelerator/IHtmlScriptureExtractor.cs
@@ -1,0 +1,9 @@
+﻿namespace SIL.Transcelerator
+{
+	public interface IHtmlScriptureExtractor
+	{
+		bool IncludeVerseNumbers { get; set; }
+
+		string GetAsHtmlFragment(int startRef, int endRef);
+	}
+}
