+++ Transcelerator/ScriptureExtractor.cs
@@ -6,13 +6,6 @@ using Paratext.PluginInterfaces;
 
 namespace SIL.Transcelerator
 {
-	public interface IHtmlScriptureExtractor
-	{
-		bool IncludeVerseNumbers { get; set; }
-
-		string GetAsHtmlFragment(int startRef, int endRef);
-	}
-
 	public class ScriptureExtractor : IHtmlScriptureExtractor
 	{
 		private readonly IProject m_project;
